pyproject.toml
[tool.poetry]
name = "jamstack-api"
version = "0.1.0"
description = "Webhook endpoints to optimize your JAMStack static site."
authors = ["Todd Birchard <toddbirchard@gmail.com>"]
maintainers = ["Todd Birchard <toddbirchard@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/toddbirchard/jamstack-api/"
repository = "https://github.com/toddbirchard/jamstack-api/"
documentation = "https://github.com/toddbirchard/jamstack-api/"
keywords = [
  "API",
  "Rest API",
  "Automation",
  "FastAPI",
  "Webhooks"
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
pydantic = "^1.8.0"
pandas = "*"
sqlalchemy = "*"
pymysql = "*"
requests = "*"
google-cloud-storage = "*"
google-cloud-bigquery = "*"
google-cloud-bigquery-storage = "*"
pillow = "*"
python-resize-image = "*"
webp-converter = "*"
python-dotenv = "*"
pyjwt="*"
mixpanel = "*"
loguru = "*"
twilio = "*"
beautifulsoup4 = "*"
extruct = "*"
simplejson = "*"
boto = "*"
botocore = "*"
uvicorn = "^0.13.0"
ddtrace = "^0.44.0"
ddtrace-asgi = "^0.3.0"
PyGithub = "^1.54"
black = "^20.8b1"

[tool.poetry.dev-dependencies]
pytest = "*"
mock = "*"
mypy = "*"
flake8 = "^3.8.4"

[tool.poetry.scripts]
run = "asgi:app"

[tool.poetry.urls]
issues = "https://github.com/toddbirchard/jamstack-api/issues"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
